Alterations to test files and armadillo after pull request review
